requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.115.5
filelock==3.13.1
fonttools==4.52.4
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.4
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
Jinja2==3.1.2
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.143
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
openai==1.54.5
orjson==3.10.11
packaging==23.2
pandas==2.1.4
peewee==3.17.5
pillow==10.2.0
platformdirs==4.2.2
propcache==0.2.0
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.2.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.16.2
rsa==4.9.1
safetensors==0.4.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.35
starlette==0.41.3
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.32.0
uvloop==0.21.0
validators==0.22.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==14.1
xgboost==2.0.3
yarl==1.17.2
yfinance==0.2.40
zipp==3.17.0
